requirements.txt
# local package
-e .

# external requirements
pytorch=1.9.0
torchvision
torchaudio
cudatoolkit=11.1
tensorboard
numpy
scipy
scikit-learn
scikit-image
matplotlib
seaborn
pandas
click
absl-py
pyyaml
jupyterlab
tqdm
sphinx
awscli
python-dotenv
duecredit
watermark
torchinfo
dominate
visdom
Pillow
jupyterlab_vim
ipywidgets
nb-clean
